refactor(api): preset last little thing
