+++ apps/api/src/common/schemas/baserResponse.ts
@@ -1,0 +1,9 @@
+import {Type} from '@sinclair/typebox';
+
+export const BaseResponse = {
+  id: Type.String({format: 'uuid'}),
+  name: Type.String(),
+  createdAt: Type.Unsafe<Date | string>({format: 'date-time'}),
+  updatedAt: Type.Unsafe<Date | string>({format: 'date-time'}),
+  ownerId: Type.String({format: 'uuid'}),
+};
