+++ apps/api/src/modules/project/schema/project.schema.ts
@@ -1,20 +1,10 @@
 import {Type} from '@sinclair/typebox';
+import {BaseResponse} from '../../../common/schemas/baserResponse';
 import {Nullable} from '../../../common/typebox/nullable';
 
-export const BaseResponse = {
-  id: Type.String({format: 'uuid'}),
-  name: Type.String(),
-  createdAt: Type.Unsafe<Date | string>({format: 'date-time'}),
-  updatedAt: Type.Unsafe<Date | string>({format: 'date-time'}),
-  ownerId: Type.String({format: 'uuid'}),
-};
-
-export const BaseProjectResponseSchema = Type.Object(
-  {...BaseResponse},
-  {
-    title: 'BaseProjectResponse',
-  },
-);
+export const BaseProjectResponseSchema = Type.Object(BaseResponse, {
+  title: 'BaseProjectResponse',
+});
 
 export const BaseSnippetEditorTabsSchema = Type.Array(
   Type.Object({
